refactor: Rename hero folder to header
