+++ src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import Footer from "../components/Footer/Footer";
-import Hero from "../components/Hero/Header";
+import Header from "../components/Header/Header";
 import MovieSearchResult from "../components/MovieSearchResult/MovieSearchResult";
 import { movies } from "../data/movies";
 import classes from "./Home.module.css";
@@ -7,7 +7,7 @@ import classes from "./Home.module.css";
 const Home: React.FC = () => {
   return (
     <div className={classes.container}>
-      <Hero />
+      <Header />
       <MovieSearchResult count={39} movies={movies} />
       <Footer />
     </div>
